requirements.txt
# Core LLM
google-generativeai>=0.8.0

# Environment
python-dotenv>=1.0.0

# Stance Firmness Classifier (DistilBERT)
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
scikit-learn>=1.3.0

# CMV Data Collection
praw>=7.7.0

# Frontend
streamlit>=1.30.0

# Utilities
numpy>=1.24.0
pandas>=1.5.0
networkx>=3.0       # Concept graph for adaptive engine

# Testing
pytest>=7.4.0
